pyproject.toml
[project]
name = "mcp-postgres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "mcp[cli]>=1.12.4",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
